// src/app/co/ada/models/DatosPersonalesEmpleado.ts
export interface DatosPersonalesEmpleado {
    id: number;
    nitsd: number;
    tipo_vinculo: number;
    nombre: string;
    tipoDocumento: string;
    nit: number;
    ciudad: string;
    email: string; 
    fechaCreacion: Date;
    estado: string;
    direccion: string;
    telefono: string;

}